refactor: 토스트 관련 패키지, 파일 정리
