+++ designsystem/src/main/java/com/mundcode/designsystem/components/toast/CustomToast.kt
@@ -1,0 +1,47 @@
+package com.mundcode.designsystem.components.toast
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.scaleIn
+import androidx.compose.animation.scaleOut
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.mundcode.designsystem.state.ToastState
+import com.mundcode.designsystem.state.rememberToastState
+import com.mundcode.designsystem.theme.CornerRadius8
+import com.mundcode.designsystem.theme.MTTextStyle
+import com.mundcode.designsystem.theme.MTToastBackground
+import com.mundcode.designsystem.theme.White
+import kotlinx.coroutines.launch
+
+
+@OptIn(ExperimentalAnimationApi::class)
+@Composable
+fun CustomToast(
+    toastState: ToastState,
+    content: @Composable () -> Unit
+) {
+    AnimatedVisibility(
+        visible = toastState.show,
+        enter = fadeIn() + scaleIn(),
+        exit = fadeOut() + scaleOut()
+    ) {
+        content()
+    }
+}
